silence static code analysis warnings about integer overflows

+++ src/ML-SNAP/compute_grid.cpp
@@ -235,8 +235,8 @@ void ComputeGrid::set_grid_local()
 
 double ComputeGrid::memory_usage()
 {
-  double nbytes = size_array_rows * size_array_cols * sizeof(double);    // grid
-  nbytes += size_array_rows * size_array_cols * sizeof(double);          // gridall
-  nbytes += size_array_cols * ngridlocal * sizeof(double);               // gridlocal
+  double nbytes = (double) size_array_rows * size_array_cols * sizeof(double);    // grid
+  nbytes += (double) size_array_rows * size_array_cols * sizeof(double);          // gridall
+  nbytes += (double) size_array_cols * ngridlocal * sizeof(double);               // gridlocal
   return nbytes;
 }
